third_party/WebKit/LayoutTests/virtual/presentation/receiver/presentation-receiver-iframe-sandboxing.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<iframe></iframe>
<script>

var iframe = document.querySelector('iframe');

async_test(t => {
  iframe.onload = t.step_func(() => {
    iframe.contentWindow.postMessage('start', '*');
  });

  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data, 'success');
  }));

  // Navigate the iframe
  iframe.src = 'resources/iframe-sandbox.html';
}, "Test that sandboxing flags apply in an iframe in a presentation receiver page.");

</script>
</body>
</html>
